Updated the manifest
Hi Jeffrey, I had to change the manifest to add the new page, as well as switching two pages that were in wrong order. It's easiest for me to make the change in branch as I think it would create a conflict. Otherwise I think your repo has all that it needs!